+++ app/src/main/java/com/example/happi/ratemycourse/LoginPageFragment.java
@@ -3,12 +3,9 @@ package com.example.happi.ratemycourse;
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.text.InputType;
-import android.text.method.DigitsKeyListener;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -17,8 +14,9 @@ import android.view.inputmethod.EditorInfo;
 import android.widget.Button;
 import android.widget.EditText;
 
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
+import com.example.happi.ratemycourse.database.UserDataHandler;
+import com.example.happi.ratemycourse.database.UserDataModel;
+import com.example.happi.ratemycourse.util.TextEncoder;
 
 
 /**
