+++ app/src/main/java/com/example/happi/ratemycourse/database/CommentsDataModel.java
@@ -1,4 +1,4 @@
-package com.example.happi.ratemycourse;
+package com.example.happi.ratemycourse.database;
 
 //import android.util.Log;
